+++ src/cli/controllers/main.js
@@ -1,4 +1,5 @@
 import { CliController } from 'cli-router';
+import Rx from 'rx';
 import Main from '../../main';
 import Saver from '../../saver';
 
@@ -6,11 +7,15 @@ class MainController extends CliController {
   index(params) {
     const main = new Main(params);
     const saver = new Saver(params);
-    return main.run()
+
+    main.run()
     .map((url) => {
       saver.save(url, params.save);
-      return null;
-    });
+      return url;
+    })
+    .subscribe();
+
+    return Rx.Observable.of(null);
   }
 }