+++ src/main/index.js
@@ -1,6 +1,7 @@
 import merge from 'lodash.merge';
 import dotenv from 'dotenv';
 import TwiterStream from 'twitter-stream-cli';
+import Formatter from '../formatter';
 
 class Main {
   constructor(opts) {
@@ -20,13 +21,20 @@ class Main {
   }
 
   run() {
-    const twiterStream = new TwiterStream({
-      track: this.opts['track-words']
-    });
-    return twiterStream.run()
-    .startWith(null)
-    .filter((x) => x !== null &&
-                   x.extended_entities &&
+    const twitterStream = new TwiterStream(this.opts);
+    const formatter = new Formatter(this.opts);
+
+    const tweetsWithImages$ = twitterStream.run()
+    // all
+    .do((x) => {
+      if (this.opts['show-all']) {
+        formatter.format([x]).forEach((item) => {
+          console.error(item);
+        });
+      }
+    })
+    // to get only tweets with images
+    .filter((x) => x.extended_entities &&
                    x.extended_entities.media &&
                    x.extended_entities.media.length > 0)
     .map((tweet) => {
@@ -34,6 +42,8 @@ class Main {
     })
     .selectMany((x, _i) => x)
     .distinct((x) => x.media_url);
+
+    return tweetsWithImages$;
   }
 }