+++ CHANGELOG.md
@@ -6,4 +6,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## v0.1.1 (2016-02-12)
 
 * Enhancements
-  - [Cli] track twiter and save to folder
+  - [Cli] track twitter and save to folder